Moved clause translator out into its own file.
