+++ src/Makefile.am
@@ -190,6 +190,8 @@ souffle_sources = \
         ast/utility/Visitor.h                              \
         ast2ram/AstToRamTranslator.cpp                     \
         ast2ram/AstToRamTranslator.h                       \
+        ast2ram/ClauseTranslator.cpp                       \
+        ast2ram/ClauseTranslator.h                         \
         ast2ram/ProvenanceClauseTranslator.cpp             \
         ast2ram/ProvenanceClauseTranslator.h               \
         interpreter/InterpreterContext.h                   \
